requirements.txt
Cython                    3.0.10
cython-bbox               0.1.5
easydict                  1.11
grad-cam                  1.5.2
h5py                      3.10.0
ipython                   8.12.3
matlab                    0.1
matplotlib                3.7.1
matplotlib-inline         0.1.7
ml-collections            0.1.1
mmcv                      2.0.0rc3
numba                     0.58.1
numpy                     1.23.5
opencv-contrib-python     4.8.0.74
opencv-python             4.8.0.74
pandas                    2.0.2
Pillow                    9.4.0
pip                       23.0.1
tensorboard               2.14.0
tensorboard-data-server   0.7.1
tensorboardX              2.6.2.2
timm                      0.9.8
tokenizers                0.13.3
tomli                     2.0.1
torch                     1.13.1
torchaudio                0.13.1
torchinfo                 1.8.0
torchsummary              1.5.1
torchtoolbox              0.1.8.2
torchvision               0.14.1
tqdm                      4.65.0
traitlets                 5.14.3
transformers              4.32.1